Dupable is now defined in terms of V @n
One immediate benefit is that the instances are much easier to write,
using the applicative instance (of course, it would have worked just
fine, if that was the only goal, to make a V2 applicative functor).

The other benefit is that there was no pleasant way to produce more
than two things from a dupable. It is possible to write a function `(a
->. (a,a)) ->. a ->. V n a`, but at this point we already have `V n`
defined. And producing a `V n a` directly is, I expect, significantly
more efficient.

Another approach would be to write a function `(a ->. (a,a)) ->. Int
->. a ->. [a]`, and rely on incomplete pattern-matchings. But this is
rather yucky.