Make {Array,Vector}.Mutable.Linear lazier
This PR changes wraps the result of `Unsafe.MutableArray.read` with an
unboxed tuple, so the array lookup can be evaluated without evaluating
the result.

Because of that, now `Data.{Array,Vector}.Mutable.Linear.read` functions
does not unnecessarily force the return value.

I added a test case about strictness issue just for `Array`'s. Because
Vector's `read` function returns the result linearly as opposed to
`Array` which wraps it with an `Unrestricted. I think this is another
error, so I'll fix that with a separate PR.

While doing that, I also noticed that the issue fixed in `0c9a436`
also applies to vectors, so this commit also fixes that while adding a
test case.

[1]: https://github.com/tweag/linear-base/commit/0c9a43692298d0c9603d9cbbda1567fcea29f431