Enrich ProbeResult type to reduce redundant reads
Previously, after obtaining a `ProbeResult`; in some cases we had to
read the returned index again. This was redundant, since `probeFrom`
already had read that index to obtain the `ProbeResult`.

This patch:

* Updates `IndexToUpdate` constructor to contain the previous value of
the to be updated cell. This saves one `read` call on `lookup`. It will
also be useful in near future when we implement an `update` function.
* Updates `IndexToSwap` constructor to contain the to be evicted
`RobinVal`. This saves us one `read` call on `tryInsertAtIndex` function.
* It also removes the redundant key field on `IndexToInsert` and
`IndexToUpdate` constructors since they always return the same key passed
in at `probeFrom` function.