Make RobinVal explicitly a sum type
Previosly, RobinVal was a type alias to a 3-tuple, where the second
and third element was only present when the first element was non-zero.
This was making the implementation error-prone; because we had to take
care to not evaluate the invalid elements when they are not present.

This patch:

* Makes `RobinVal` a data type instead of a type-alias.
* Wraps the `RobinVal` with a `Maybe`, so they are not partial.

While doing that, it also replaces a meant-to-be-internal `Show` instance
with a `_debugShow` function; to hide it from the API (there is no way
to hide an instance).