Remove redundant information on HashMap's constructor
HashMap's constructor was storing the size of the underlying array
in its constructor. Therefore it had to do some book-keeping to make
it consistent when doing resizes and copies.

However, that information is readily available from the array. So,
this patch removes that field from the constructor and replaces with an
accessor function called "capacity", which uses `Array.length` underneath.

After the change, the `Size` newtype was used only in two places. I
assumed its purpose was to distinguish the "size" field from the "count"
field right next to it. Since this is not the case anymore, this patch
also removes the newtype.