Implement Data.List.Linear
This commit adds linear versions of the majority of Data.List functions.

In order to do that, it also adds `Consumable` instances for `NonEmpty`
and `Either`.

It causes a minor interface change, from

```
foldl :: (b #-> a -> b) -> b #-> [a] -> b
```

to

```
foldl :: (b #-> a #-> b) -> b #-> [a] #-> b
```

Which is not a major issue since the former can be derived from latter
by wrapping list elements with `Ur`.