Internal HashMap improvements (#176)
* Enrich ProbeResult type to reduce redundant reads

Previously, after obtaining a `ProbeResult`; in some cases we had to
read the returned index again. This was redundant, since `probeFrom`
already had read that index to obtain the `ProbeResult`.

This patch:

* Updates `IndexToUpdate` constructor to contain the previous value of
the to be updated cell. This saves one `read` call on `lookup`. It will
also be useful in near future when we implement an `update` function.
* Updates `IndexToSwap` constructor to contain the to be evicted
`RobinVal`. This saves us one `read` call on `tryInsertAtIndex` function.
* It also removes the redundant key field on `IndexToInsert` and
`IndexToUpdate` constructors since they always return the same key passed
in at `probeFrom` function.

* Make RobinVal explicitly a sum type

Previosly, RobinVal was a type alias to a 3-tuple, where the second
and third element was only present when the first element was non-zero.
This was making the implementation error-prone; because we had to take
care to not evaluate the invalid elements when they are not present.

This patch:

* Makes `RobinVal` a data type instead of a type-alias.
* Wraps the `RobinVal` with a `Maybe`, so they are not partial.

While doing that, it also replaces a meant-to-be-internal `Show` instance
with a `_debugShow` function; to hide it from the API (there is no way
to hide an instance).

* Remove redundant information on HashMap's constructor

HashMap's constructor was storing the size of the underlying array
in its constructor. Therefore it had to do some book-keeping to make
it consistent when doing resizes and copies.

However, that information is readily available from the array. So,
this patch removes that field from the constructor and replaces with an
accessor function called "capacity", which uses `Array.length` underneath.

After the change, the `Size` newtype was used only in two places. I
assumed its purpose was to distinguish the "size" field from the "count"
field right next to it. Since this is not the case anymore, this patch
also removes the newtype.

* Improve HashMap documentation