Change the parameter order for optics
For reasons unknown, and almost certainly my fault, the order of the
type arguments of optic types (lens, traversal, …) was not the same as
in the lens library.

Even if I had a reason at the time (though I suspect it's just a
mistake), it couldn't have been strong enough to overcome the
principle of least surprise of being compatible with a well-used
package.