Add a Github action ci
Now that we don't need to build GHC ourselves for the sake of
linear-base to work, we don't need to lock a super beefy runner for 1h
from time to time.

On the other hand, the buildkite runner didn't run ci for
contributions outside of Tweag. Having a Github-action based ci will
be better for contributors.