Put Unsafe.Linear back in place
The contents of `Unsafe.Linear` were moved to
`Unsafe.Linear.Internal`, presumably to avoid cyclical imports
with `Generically` stuff. `Generically` is no longer unsafe,
so we don't need that, at least for now. We can get it back
if we want/need, but it doesn't belong in this PR and it makes
the diff hard for me to use.