A more robust implementation of Replicator.elim
I rely on more type families (I resurrected the `FunN` type family,
for instance) and additional type classes in order to give robust
hints to the type checker to infer some of the arguments based on the
others in multiple possible ways.

The previous version was abusing functional dependencies to that
effect. The problem is that it almost certainly depended on bugs. Plus
some behaviour changed in 9.2 making the recursive instance not
typecheck anymore.

The actual (staging) recursion now uses (type-level) Peano integers,
this way we can get rid of the `OVERLAPPABLE` instances altogether,
and avoid the sort of fragile behaviour that caused the failure in
9.2.

The 'elim' function itself has more type arguments, so I took it out
of the type class. A collection of type families and type classes fill
the role that functional dependencies used to play: figuring out `f`
from `n` or `n` from `f`.

Fixes #388.