Don't use deprecated `testProperty` from tasty-hedgehog
The `testProperty` function has been deprecated in the current 1.2
version. See https://github.com/qfpl/tasty-hedgehog/pull/42 for the
reason.

The `testProperty` function is replaced by `testPropertyNamed` which
requires one extra argument (the name of the test function as a
string).

There were a test for which the property didn't have a name, so I had
to name it. We also need `-XOverloadedString` in every module (some
already have it).