Array benchmarks: use same implementation for both data structures
I simply introduced a type class to abstract over the differences. I
almost didn't do it because I was afraid that the extra `Ur` and `(,)`
would pollute the result and bias the result in our favours. Maybe it
does. The reads are a bit less different in this style. I guess it was
to be expected because reads are so fast. But the benchmark still says
that our implementation is slow. So I say: good enough.